Remove Code duplication, fix #674
