+++ app/Http/Resources/UserResource.php
@@ -2,37 +2,21 @@
 
 namespace App\Http\Resources;
 
-use App\Http\Controllers\Backend\User\ProfilePictureController;
 use Illuminate\Http\Request;
 use Illuminate\Http\Resources\Json\JsonResource;
 
 class UserResource extends JsonResource
 {
+    protected bool $UserResource = true;
+
     /**
      * Transform the resource into an array.
      *
      * @param Request $request
      *
-     * @return array
+     * @return UserBaseResource
      */
-    public function toArray($request): array {
-        return [
-            'id'                => (int) $this->id,
-            'displayName'       => (string) $this->name,
-            'username'          => (string) $this->username,
-            'profilePicture'    => ProfilePictureController::getUrlForUserId($this->id),
-            'trainDistance'     => (float) $this->train_distance,
-            'trainDuration'     => (int) $this->train_duration,
-            'trainSpeed'        => (float) $this->averageSpeed,
-            'points'            => (int) $this->points,
-            'twitterUrl'        => $this->twitterUrl ?? null,
-            'mastodonUrl'       => $this->mastodonUrl ?? null,
-            'privateProfile'    => (bool) $this->private_profile,
-            'userInvisibleToMe' => (bool) $this->userInvisibleToMe,
-            'muted'             => (bool) $this->muted,
-            'following'         => (bool) $this->following,
-            'followPending'     => (bool) $this->followPending,
-            'preventIndex'      => (bool) $this->prevent_index,
-        ];
+    public function toArray($request): UserBaseResource {
+        return new UserBaseResource($this);
     }
 }
